+++ CHANGELOG.md
@@ -21,6 +21,7 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 - locales for table
 - complete table for orders, with proper UTC search
 - remembering last selection from data selection menu
+- remembering last chosen workplace from data selection
 
 ## [2021.1.2.24] - 2021-02-24